+++ src/lib/services/nft/sequencer.ts
@@ -19,6 +19,7 @@ import {
   zNftsByAccountResponseSequencer,
   zNftsResponseSequencer,
 } from "../types";
+import { getArchivalEndpoint } from "../utils";
 
 export const getNftsSequencerLoop = async (
   endpoint: string,
@@ -55,8 +56,11 @@ export const getNftsSequencer = async (
   paginationKey: Option<string>,
   limit: number
 ) => {
+  // TODO: remove this when backend fix the stagesync issue
+  const archivalEndpoint = getArchivalEndpoint(endpoint, endpoint);
+
   const { data } = await axios.get(
-    `${endpoint}/indexer/nft/v1/tokens/by_collection/${collectionAddress}`,
+    `${archivalEndpoint}/indexer/nft/v1/tokens/by_collection/${collectionAddress}`,
     {
       params: {
         "pagination.key": paginationKey,