+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Improvements
 
 - [#1318](https://github.com/alleslabs/celatone-frontend/pull/1318) Update ESLint rules
+- [#1327](https://github.com/alleslabs/celatone-frontend/pull/1327) Support show scrollbar when table overflow
 
 ### Bug fixes
 
@@ -53,6 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1326](https://github.com/alleslabs/celatone-frontend/pull/1326) Add fallback JSON RPC endpoint for evm tx details
+- [#1324](https://github.com/alleslabs/celatone-frontend/pull/1324) Update initia widget, utils to v1.6.3
 - [#1316](https://github.com/alleslabs/celatone-frontend/pull/1316) Fix select input disabled option, add EVM verification entry
 - [#1301](https://github.com/alleslabs/celatone-frontend/pull/1301) Upgrade Nextjs 14 to 15
 - [#1294](https://github.com/alleslabs/celatone-frontend/pull/1294) Upgrade React & ESLint version
@@ -61,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bug fixes
 
+- [#1323](https://github.com/alleslabs/celatone-frontend/pull/1323) Update collection, creator data type to nullable
 - [#1322](https://github.com/alleslabs/celatone-frontend/pull/1322) Fix archival fallback logic for no data case
 - [#1319](https://github.com/alleslabs/celatone-frontend/pull/1319) Fix table layout in markdown, fix badges layout in proposal text
 - [#1314](https://github.com/alleslabs/celatone-frontend/pull/1314) Fix css, sentence case in EVM verification flow